auth-server/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib
python-dotenv